+++ modules/azure/aks/variables.tf
@@ -13,6 +13,11 @@ variable "name" {
   type        = string
 }
 
+variable "unique_suffix" {
+  description = "Unique suffix that is used in globally unique resources names"
+  type        = string
+}
+
 variable "core_name" {
   description = "The commonName for the core infrastructure"
   type        = string
@@ -155,4 +160,10 @@ variable "azure_metrics_identity" {
     id           = string
     principal_id = string
   })
+}
+
+variable "aks_audit_log_retention" {
+  description = "The aks audit log retention in days, 0 = infinite"
+  type        = number
+  default     = 180
 }